Remove sinon fake timer implementation.
Sinon's fake timer implementation monkey patches globals. This
is needed because previously the Prober implementation depended
on the global state of Date.now()

The prober has been changed to have an arbitrary `now` function
passed in (defaults to `Date.now()`). The test has been rewritten
to pass in a fake timer with a now() method that returns a fake
time.